feat: extract utility functions for formatting time and downloads into a new module
