+++ app/src/widgets/card.rs
@@ -3,8 +3,8 @@
  * All rights reserved.
  */
 
+use crate::utils::Utils;
 use egui::{Response, TextureHandle, Ui, Widget};
-use web_sys::js_sys::Date;
 
 pub struct CardWidget {
     pub thumbnail: Option<TextureHandle>,
@@ -15,34 +15,6 @@ pub struct CardWidget {
     pub downloads: u32,
     pub rating: f32,
     pub last_updated: f64,
-}
-
-impl CardWidget {
-    pub fn format_time_elapsed(last_updated: f64) -> String {
-        let elapsed_seconds = (Date::now() - last_updated) / 1000.0;
-        let time_ranges = [
-            (60.0, "seconds", 1.0),
-            (3600.0, "minutes", 60.0),
-            (86400.0, "hours", 3600.0),
-            (31536000.0, "days", 86400.0),
-            (31536000.0, "months", 2592000.0),
-            (f64::INFINITY, "years", 31536000.0),
-        ];
-        for &(limit, unit, divisor) in &time_ranges {
-            if elapsed_seconds < limit {
-                return format!("⏳ updated {:.0} {} ago", elapsed_seconds / divisor, unit);
-            }
-        }
-        unreachable!("Time ranges should cover all cases");
-    }
-
-    pub fn format_downloads(downloads: u32) -> String {
-        match downloads {
-            d if d >= 1_000_000 => format!("{:.1}M", d as f64 / 1_000_000.0),
-            d if d >= 1_000 => format!("{:.1}K", d as f64 / 1_000.0),
-            _ => downloads.to_string(),
-        }
-    }
 }
 
 impl Widget for CardWidget {
@@ -89,9 +61,9 @@ impl Widget for CardWidget {
                     ui.label(&self.description);
                     ui.label(&self.platform);
                     ui.with_layout(egui::Layout::right_to_left(egui::Align::Center), |ui| {
-                        ui.label(Self::format_time_elapsed(self.last_updated));
+                        ui.label(Utils::format_time_elapsed(self.last_updated));
                         ui.label(format!("★ {:.1}", self.rating.clamp(0.0, 10.0)));
-                        ui.label(format!("📥 {}", Self::format_downloads(self.downloads)));
+                        ui.label(format!("📥 {}", Utils::format_downloads(self.downloads)));
                     });
                 });
             })
